spawner (zombie siege/phantom)
