+++ src/main/java/ovh/corail/tombstone/spawner/package-info.java
@@ -1,0 +1,3 @@
+@javax.annotation.ParametersAreNonnullByDefault
+@mcp.MethodsReturnNonnullByDefault
+package ovh.corail.tombstone.spawner;
