+++ src/main/resources/META-INF/accesstransformer.cfg
@@ -18,6 +18,5 @@ public net.minecraft.server.MinecraftServer field_110456_c # serverProxy
 public net.minecraft.loot.conditions.LootConditionManager func_237475_a_(Ljava/lang/String;Lnet/minecraft/loot/ILootSerializer;)Lnet/minecraft/loot/LootConditionType; # register
 public net.minecraft.util.text.Color field_240740_c_ # int color
 public-f net.minecraft.world.server.ServerWorld field_241104_N_ # special spawners
-public-f net.minecraft.advancements.FunctionManager field_193070_d # functions
-public-f net.minecraft.world.gen.OverworldChunkGenerator field_225495_n # villageSiege
-public-f net.minecraft.world.gen.OverworldChunkGenerator field_203230_r # phantomSpawner
+public net.minecraft.world.spawner.PhantomSpawner field_203233_a # ticksUntilSpawn
+#public-f net.minecraft.advancements.FunctionManager field_193070_d # functions
